frame,transport: add support for named bind markers
This commit adds support for CQL named bind markers,
i.e. markers in a form of ":name_here" instead of customary "?".
Named markers are not recommended due to the fact that they induce
more overhead than unnamed ones, but it's still part of the CQL
protocol, so we should support them.

Fixes #360